Delete useless ugly contructor

+++ boolfun/headers/polynom.h
@@ -24,7 +24,7 @@ namespace bf {
 			void set_coeff (uint32_t index, bvect32 coeff);
 
 			bvect32	 get_coeff (uint32_t index);
-			uint32_t get_length ();
+			uint64_t get_length ();
 			GF* get_field ();
 			
 			unsigned int get_input_length ();
@@ -33,7 +33,7 @@ namespace bf {
 
 		private:
 			std::vector<bvect32> _coeff;
-			uint32_t _length;
+			uint64_t _length;
 			GF* _field;
 			unsigned int _n; //input
 			unsigned int _m;
